modify organizations, constructionObjects, departments

+++ src/core/departments/departments.service.ts
@@ -1,45 +1,49 @@
-import { Injectable } from "@nestjs/common";
+import {
+    BadRequestException,
+    Injectable,
+    InternalServerErrorException,
+} from "@nestjs/common";
 import { InjectModel } from "@nestjs/sequelize";
-import { Transaction } from "sequelize";
-import { CreateDepartmentDto } from "./dto/create-department.dto";
-import { UpdateDepartmentDto } from "./dto/update-department.dto";
+import { IncludeOptions, Transaction } from "sequelize";
+import { Sequelize } from "sequelize-typescript";
+import { Employee } from "../employees/entities/employee.entity";
+import { Organization } from "../organizations/entities/organization.entity";
+import { CreateDepartmentDto } from "./dto/createDepartment.dto";
+import { UpdateDepartmentDto } from "./dto/updateDepartment.dto";
 import { Department } from "./entities/department.entity";
 
 @Injectable()
 export class DepartmentsService {
+    attributes: Array<string>;
+    include: Array<IncludeOptions>;
+
     constructor(
         @InjectModel(Department)
         private departmentsRepository: typeof Department,
-    ) {}
+        private sequelize: Sequelize,
+    ) {
+        // Параметры запросов к БД
+        this.attributes = ["id", "name"];
+        this.include = [
+            { model: Employee, required: false },
+            { model: Organization, required: false },
+        ];
+    }
 
     async create(
         createDepartmentDto: CreateDepartmentDto,
         transaction?: Transaction,
     ) {
-        return await this.departmentsRepository.create(createDepartmentDto, {
-            transaction,
-        });
-    }
-
-    async findAll(limit?: number, offset?: number, transaction?: Transaction) {
-        if (!limit || !offset) {
-            return await this.departmentsRepository.findAndCountAll({
-                transaction,
-            });
-        } else {
-            return await this.departmentsRepository.findAndCountAll({
-                limit,
-                offset,
-                transaction,
-            });
+        try {
+            return await this.departmentsRepository.create(
+                createDepartmentDto,
+                {
+                    transaction,
+                },
+            );
+        } catch (e) {
+            throw new InternalServerErrorException(e);
         }
-    }
-
-    async findOne(id: number, transaction?: Transaction) {
-        return await this.departmentsRepository.findOne({
-            where: { id },
-            transaction,
-        });
     }
 
     async update(
@@ -47,16 +51,82 @@ export class DepartmentsService {
         updateDepartmentDto: UpdateDepartmentDto,
         transaction?: Transaction,
     ) {
-        return await this.departmentsRepository.update(updateDepartmentDto, {
-            where: { id },
-            transaction,
-        });
+        try {
+            const candidate = await this.findOne(id, transaction);
+
+            if (candidate) {
+                return await this.departmentsRepository.update(
+                    updateDepartmentDto,
+                    {
+                        where: { id },
+                        transaction,
+                    },
+                );
+            }
+        } catch (e) {
+            throw e;
+        }
+    }
+
+    async findAll(limit?: number, offset?: number, transaction?: Transaction) {
+        try {
+            if (!limit || !offset) {
+                return await this.departmentsRepository.findAndCountAll({
+                    attributes: this.attributes,
+                    include: this.include,
+                    order: [["name", "ASC"]],
+                    distinct: true,
+                    transaction,
+                });
+            } else {
+                return await this.departmentsRepository.findAndCountAll({
+                    limit,
+                    offset,
+                    attributes: this.attributes,
+                    include: this.include,
+                    order: [["name", "ASC"]],
+                    distinct: true,
+                    transaction,
+                });
+            }
+        } catch (e) {
+            throw e;
+        }
+    }
+
+    async findOne(id: number, transaction?: Transaction) {
+        try {
+            const candidate = await this.departmentsRepository.findByPk(id, {
+                attributes: this.attributes,
+                include: this.include,
+                order: [["name", "ASC"]],
+                transaction,
+            });
+
+            if (!candidate) {
+                throw new BadRequestException("Участок не найден");
+            }
+
+            return candidate;
+        } catch (e) {
+            throw e;
+        }
     }
 
     async remove(id: number, transaction?: Transaction) {
-        return await this.departmentsRepository.destroy({
-            where: { id },
-            transaction,
-        });
+        try {
+            const candidate = await this.findOne(id, transaction);
+
+            if (!candidate) {
+                throw new BadRequestException("Участок не найден");
+            }
+
+            return await this.departmentsRepository.destroy({
+                where: { id },
+                transaction,
+            });
+        } catch (e) {
+            throw e;
+        }
     }
 }
